chore: Update testing setup to use Jest (#9)
* uninstall jasmine/karma packages

* chore: Update angular.json to remove karma test builder

* chore: Remove karma test builder and jasmine/karma packages

This commit removes the karma test builder from the angular.json file and uninstalls the jasmine/karma packages. These changes are made to streamline the testing setup and remove unnecessary dependencies.

* chore: Install the jest package in your project

This commit updates the npm dependencies in the package.json file. Specifically, it adds the "@types/jest" and "jest-preset-angular" packages to enable Jest testing in the project. This update aligns with recent changes made to streamline the testing setup and remove unnecessary dependencies.

* chore: Add setup-jest.ts file for Jest testing

This commit adds a new file, setup-jest.ts, to the ClientApp/src directory. The file imports the 'jest-preset-angular/setup-jest' package, which is necessary for Jest testing in the project. This change aligns with recent updates to streamline the testing setup and remove unnecessary dependencies.

* chore: Update tsconfig.spec.json for Jest testing

The tsconfig.spec.json file has been updated to include the "jest" type in the compilerOptions. This change aligns with recent updates to streamline the testing setup and remove unnecessary dependencies.

* chore: Update Jest configuration for testing

This commit updates the Jest configuration in the package.json file to include new test scripts and setup files. It adds the "jest" package as a dependency and configures Jest to use the "jest-preset-angular" preset. Additionally, it adds the "test:watch" and "test:coverage" scripts for running tests in watch mode and generating test coverage reports. These changes align with recent updates to streamline the testing setup and remove unnecessary dependencies.